Refactor 'backends' to 'interpolate'
