--- sarracen/backends/__init__.py
@@ -1,4 +1,0 @@
-from sarracen.backends.base_backend import BaseBackend
-from sarracen.backends.cpu_backend import CPUBackend
-from sarracen.backends.gpu_backend import GPUBackend
-from sarracen.backends.cupy_backend import CuPyBackend
