+++ sarracen/interpolate/__init__.py
@@ -1,0 +1,5 @@
+from sarracen.interpolate.base_backend import BaseBackend
+from sarracen.interpolate.cpu_backend import CPUBackend
+from sarracen.interpolate.gpu_backend import GPUBackend
+from sarracen.interpolate.interpolate import interpolate_2d_cross, interpolate_2d, interpolate_3d,\
+    interpolate_3d_cross, interpolate_3d_vec, interpolate_3d_cross_vec, interpolate_2d_vec
